Include first_pts=0 when resampling
Switching to ffmpeg-python with filter('aresample', ...) introduced a
subtle bug. The old code was using `aresample={rate}:first_pts=0`, while
the new code left off the `first_pts` part.

This causes files with a PTS offset to read with an incorrect length.
This meant that the raw PCM data could either be missing or have extra
samples at the front.